Fix clang warnings

+++ sway/commands.c
@@ -870,8 +870,8 @@ static struct cmd_results *cmd_move(int argc, char **argv) {
 			uint32_t w = size->w - g.size.w;
 			uint32_t h = size->h - g.size.h;
 
-			view->x = g.origin.x = MIN(w, MAX(x, 0));
-			view->y = g.origin.y = MIN(h, MAX(y, 0));
+			view->x = g.origin.x = MIN((int32_t)w, MAX(x, 0));
+			view->y = g.origin.y = MIN((int32_t)h, MAX(y, 0));
 
 			wlc_view_set_geometry(view->handle, 0, &g);
 		}
